Reorganize classes and implement extensions
